+++ site/src/components/Prose.jsx
@@ -17,7 +17,9 @@ export function Prose({ as: Component = 'div', className, ...props }) {
         // pre
         'prose-pre:rounded-xl prose-pre:bg-slate-900 prose-pre:shadow-lg dark:prose-pre:bg-slate-800/60 dark:prose-pre:shadow-none dark:prose-pre:ring-1 dark:prose-pre:ring-slate-300/10',
         // hr
-        'dark:prose-hr:border-slate-800'
+        'dark:prose-hr:border-slate-800',
+        // strong/bold text
+        'prose-strong:text-slate-900 dark:prose-strong:text-slate-200 prose-strong:font-semibold'
       )}
       {...props}
     />